Move sig-storage tests to storage subdir
